backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["송형주 <hyungjoo.song@sktelecom.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
langgraph = "^0.2.35"
langchain = "^0.3.3"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.2"
langchain-openai = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
